chatgpt/requirements.txt
lint
black
pytest
pytest-cov
click
ipython
ipykernel
openai==0.27.4
numpy
pandas
tiktoken
typing
textract
langchain
pinecone-client[grpc]
termcolor
streamlit